move FrameBuffer cache to Renderer
